Termcap functionality beginning
Basic movement left and right is possible, but line editing is still not
implemented. Backspace is implemented however.

It's worth looking into just how effective termcaps is, and whether it is
possible to flush the entire contents of termcaps output into a
character string - if so, then there will be no need to regularly
reallocate and reprint a single line representing the shell command.