CTRL-ARROW motions now functional
For whatever reason, CTRL-UP and CTRL-DOWN are interpreted as the same
key in my working environment, so for the time being I have used
SHIFT-UP and SHIFT-DOWN to move directly up and down rows. UP and DOWN
have been saved for scrolling through history (if we ever want to
implement this).

CTRL-LEFT and CTRL-RIGHT will now jump to the first letter of the
previous word and the next word respectively. This functionality applies
also for multiple lines of input.