Attempting to recreate tokenizer w/out linked list
Tokenization should be possible using only the original line of input -
by separation into tokens by placing NULL bytes into the delimited areas
of the input line, and storing tokens as pointers to the beginning of
each section.

This would be relatively simple without the functionality of the
backtick character '`', which opens a subshell to run whatever command
is withing the backtick section, and then uses this output as a part of
the original command. As subshell output is of unknown length, it
prevents this delimiting method of token generation - lending support to
a linked list implementation.