Commit while in the middle of new lexer cleanup
Currently working to tidy the code for the new lexer into a shorter, and
more readable form. Should be relatively simple to complete.

More testing still required to ensure that token generation is exactly
what is required.

Backtick functionality is desirable, but must wait until almost
everything else has been completed. Will require creation of a subshell
- using the execute_cmd function of minishell should be fine as long as
output is formatted into a string - for placement into the parent
command. Otherwise, reading of standard output could be necessary and
that might be difficult.