Codingame finished - returning to minishell
I can't remember exactly where I was in this - but according to the git
changes the last files I was working on were:

- finitestatemachine.c
- inputconcat.c
- lexutils.c

And modified:

- ryantokens.c

I'm pretty sure I was working to make all sorts of input viable, such
as:

- cat < file_name - There should only be two tokens for this, with the
'file_name' token being assigned a type of INPUTFILE or something, to
hand it's contents over to the previous command. Will have to test what
a line like "cat < file_name_1 < file_name_2" does.

- echo $HELLO - I think this is currently working to an extent, the
environment variable is given a token ENV, which will change it during
processing. However, something needs to be done for:
	- $VAR hello, where VAR = "echo", the environment token should
	become a command token, with the relevant value.
	- echo "$HOME", works just like echo $HOME
	- echo '$HOME' does not expand the environment variable

Perhaps a function token_processing() is required to deal with these
special cases, for example, checking to see if a "COMMAND" token is in
the format of an environment variable, and expanding it if so.

Simple to-do list:

	- Deal with "<" symbol
	- Deal with "<<" (or ">>"), I can't remember which is required
	- Deal with environment variables in weird places, like as the
	command, or inside a string literal (''').