Fixed a couple input things
Inserting into the line was not working AT ALL, all inserted characters
were simply being appended to the end of the input string. Fixed.

When inserting into the first line of input, and there were multiple
lines of input, with the last line being in the final row of the
terminal, scrolling would not take place, resulting in some of the input
line not appearing. Fixed now by adding a static scroll_lines() function
in src/add_char.c.