Improved line editing further
An arrow-key press immediately after pasting from clipboard could result
in the character value of that key press being output to the terminal -
this is now prevented with an `if` statement in char_input() (in
input1.c). Special commands such as CTRL-U also now work immediately
after pasting.

sleep(5) left behind from testing removed from src/gen_tokens.c.

After the escape key is pressed, the terminal waits to see what input
will be next, to determine if an FN is desired, or if the keypress is
the ESC key alone. This was causing an unnacceptable delay in ESC-delete
functionality - fixed by checking immediately if the second char in
input buffer is 127 for backspace ([27, 127, 0, 0, 0, 0] representing
ESC-backspace).
ESC-backspace will delete an entire word from input.