Implemented multiline editing
Some testing to ensure functionality is recommended though.

Many files changed to ensure functionality when multiline is active -
mostly using a new part of the t_line structure called multiline_len,
which stores the length of the string contained prior to the entry of
multi-line editing.

This portion of the input string is off limits to editing.