One bug fixed
"echo $HOM%R" was segfaulting when using address sanitizer. Due to
problematic code in evaluate_env(), that could result in going out of
bounds at some points.

A null is no longer inserted into the line by default when an
environment variable that does not exist is called. However null should
be inserted by gen_tokens() anyway. There may be edge cases that this
fix breaks, so more testing is required to make sure this is not the
case.