Implemented multiline for "\"
Solution here was a bit hacky, and as a result CTRL-D while waiting for
a \ on multiline will give a slight incorrect error message (having a
trailing \ will put the lexer into the INDOUBLEQUOTE state), but
otherwise functionality looks okay.

Also removed some code which meant that "\ " was not being interpreted
as a new token, when a single space can be a valid token as long as it
is escaped.

Hoping these implementations haven't broken other things...