Fixed `echo this > >` crash
`echo this > > ` would crash IF it was the first line of input, due to
vec_destroy being called on an uninitialized vector. Now protected by an
`if` statement.

`echo this >>` no longer outputs "this", now it detects a parse error, as
there is no file to append to.

A similar fix has been put in place for lines like `echo this >;`, and
similar for an unexpected pipe symbol.